// client/src/types.ts
export interface Place {
  id: number
  trip_id: number
  name: string
  description: string | null
  lat: number | null
  lng: number | null
  address: string | null
  category_id: number | null
  icon: string | null
  price: string | null
  image_url: string | null
  google_place_id: string | null
  osm_id: string | null
  place_time: string | null
  end_time: string | null
  created_at: string
}